rework file locations
