+++ backend/src/routes/imageRoutes.ts
@@ -1,6 +1,6 @@
 import { param } from "express-validator";
+import { RouteDefinition } from "../common/types.js";
 import { getImage } from "../controllers/imageController.js";
-import { RouteDefinition } from "./routes.js";
 
 export const ImageRoutes: RouteDefinition = {
     path: '/image/:image_id',
